Refactored some data related to the challenge state.
