+++ src/main/java/dev/skippaddin/allAndOnlyChests/commands/StructureCommand.java
@@ -1,6 +1,7 @@
 package dev.skippaddin.allAndOnlyChests.commands;
 
 import dev.skippaddin.allAndOnlyChests.AllAndOnlyChests;
+import dev.skippaddin.allAndOnlyChests.challenge.ChallengeData;
 import dev.skippaddin.allAndOnlyChests.scoreboard.StructureScoreboard;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TextComponent;
@@ -21,14 +22,14 @@ public class StructureCommand implements CommandExecutor {
                 if (args[0].equalsIgnoreCase("finish")) {
                     if (p.hasPermission("AllAndOnlyChests.command.structure.finish")) {
                         if (!AllAndOnlyChests.getSelectedStructure().isEmpty()) {
-                            AllAndOnlyChests.getStructureProgress().replace(AllAndOnlyChests.getSelectedStructure(), true);
+                            ChallengeData.getStructureProgress().replace(AllAndOnlyChests.getSelectedStructure(), true);
                             AllAndOnlyChests.setSelectedStructure("");
                             TextComponent textComponent = Component.text("Structure completed!", NamedTextColor.GOLD);
                             p.sendMessage(textComponent);
                             p.playSound(p, Sound.UI_TOAST_CHALLENGE_COMPLETE, 1f, 1f);
                             StructureScoreboard scoreboard = StructureScoreboard.getInstance();
                             scoreboard.complete();
-                            AllAndOnlyChests.setSaved(false);
+                            ChallengeData.setSaved(false);
                         } else {
                             TextComponent textComponent = Component.text("No structure selected!", NamedTextColor.RED);
                             p.sendMessage(textComponent);
