+++ src/main/java/dev/skippaddin/allAndOnlyChests/listeners/ItemDropListener.java
@@ -1,7 +1,7 @@
 package dev.skippaddin.allAndOnlyChests.listeners;
 
 import com.destroystokyo.paper.event.block.BlockDestroyEvent;
-import dev.skippaddin.allAndOnlyChests.AllAndOnlyChests;
+import dev.skippaddin.allAndOnlyChests.challenge.ChallengeData;
 import org.bukkit.Material;
 import org.bukkit.block.*;
 import org.bukkit.entity.Item;
@@ -24,23 +24,23 @@ public class ItemDropListener implements Listener {
     //Prevents getting items from item frame
     @EventHandler
     public void onItemFrameDamage(EntityDamageByEntityEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed() && e.getEntity() instanceof ItemFrame) {
+        if (!ChallengeData.isDropsAllowed() && e.getEntity() instanceof ItemFrame) {
             e.setCancelled(true);
         }
     }
 
     @EventHandler
     public void onDestroyBlock(BlockDestroyEvent e) {
-        if (!AllAndOnlyChests.getPlacedBlocks().remove(e.getBlock()) && !AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.getPlacedBlocks().remove(e.getBlock()) && !ChallengeData.isDropsAllowed()) {
             e.setWillDrop(false);
         } else {
-            AllAndOnlyChests.setSaved(false);
+            ChallengeData.setSaved(false);
         }
     }
 
     @EventHandler
     public void onHangingBreak(HangingBreakEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             e.setCancelled(true);
         }
     }
@@ -48,10 +48,10 @@ public class ItemDropListener implements Listener {
     @EventHandler
     public void onBlockBreak(BlockBreakEvent e) {
         Block block = e.getBlock();
-        if (!AllAndOnlyChests.getPlacedBlocks().remove(block) && !AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.getPlacedBlocks().remove(block) && !ChallengeData.isDropsAllowed()) {
             block.setType(Material.AIR);
         } else {
-            AllAndOnlyChests.setSaved(false);
+            ChallengeData.setSaved(false);
         }
     }
 
@@ -61,46 +61,46 @@ public class ItemDropListener implements Listener {
         if (blockState instanceof InventoryHolder && !(blockState instanceof Dispenser || blockState instanceof DecoratedPot)) {
             for (Item item : e.getItems()) {
                 ItemStack itemStack = item.getItemStack();
-                if (itemStack.hasItemMeta() && AllAndOnlyChests.getStructureMap().contains(itemStack.getItemMeta().getItemName())) {
+                if (itemStack.hasItemMeta() && ChallengeData.getStructureSet().contains(itemStack.getItemMeta().getItemName())) {
                     e.setCancelled(true);
                     break;
                 }
             }
-        } else if (!AllAndOnlyChests.isDropsAllowed() && (blockState instanceof Dispenser || blockState instanceof DecoratedPot)) {
+        } else if (!ChallengeData.isDropsAllowed() && (blockState instanceof Dispenser || blockState instanceof DecoratedPot)) {
             e.setCancelled(true);
         }
     }
 
     @EventHandler
     public void onBlockPlace(BlockPlaceEvent e) {
-        AllAndOnlyChests.getPlacedBlocks().add(e.getBlockPlaced());
-        AllAndOnlyChests.setSaved(false);
+        ChallengeData.getPlacedBlocks().add(e.getBlockPlaced());
+        ChallengeData.setSaved(false);
     }
 
     @EventHandler
     public void onEntityDropItem(EntityDropItemEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             e.setCancelled(true);
         }
     }
 
     @EventHandler
     public void onEntityDeath(EntityDeathEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             e.getDrops().clear();
         }
     }
 
     @EventHandler
     public void onCatch(PlayerFishEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed() && (e.getState() == PlayerFishEvent.State.CAUGHT_ENTITY || e.getState() == PlayerFishEvent.State.CAUGHT_FISH)) {
+        if (!ChallengeData.isDropsAllowed() && (e.getState() == PlayerFishEvent.State.CAUGHT_ENTITY || e.getState() == PlayerFishEvent.State.CAUGHT_FISH)) {
             e.getCaught().remove();
         }
     }
 
     @EventHandler
     public void onExplodeEvent(BlockExplodeEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             for (Block block : e.blockList()) {
                 block.setType(Material.AIR);
             }
@@ -109,7 +109,7 @@ public class ItemDropListener implements Listener {
 
     @EventHandler
     public void onEntityExplode(EntityExplodeEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             for (Block block : e.blockList()) {
                 block.setType(Material.AIR);
             }
@@ -119,23 +119,23 @@ public class ItemDropListener implements Listener {
 
     @EventHandler
     public void onCrafterCraftItem(CrafterCraftEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed() && !AllAndOnlyChests.getPlacedBlocks().contains(e.getBlock())) {
+        if (!ChallengeData.isDropsAllowed() && !ChallengeData.getPlacedBlocks().contains(e.getBlock())) {
             e.setCancelled(true);
         }
     }
 
     @EventHandler
     public void onLeaveDecay(LeavesDecayEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed()) {
+        if (!ChallengeData.isDropsAllowed()) {
             e.getBlock().setType(Material.AIR);
         }
     }
 
     @EventHandler
     public void onPlayerInteract(PlayerInteractEvent e) {
-        if (!AllAndOnlyChests.isDropsAllowed() && e.getAction() == Action.RIGHT_CLICK_BLOCK) {
+        if (!ChallengeData.isDropsAllowed() && e.getAction() == Action.RIGHT_CLICK_BLOCK) {
             Block block = e.getClickedBlock();
-            if (block != null && block.getType() == Material.CHISELED_BOOKSHELF && !AllAndOnlyChests.getPlacedBlocks().contains(block)) {
+            if (block != null && block.getType() == Material.CHISELED_BOOKSHELF && !ChallengeData.getPlacedBlocks().contains(block)) {
                 e.setCancelled(true);
             }
         }
